backend/requirements.txt
Flask>=2.0.0
PyMuPDF>=1.18.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
spacy>=3.0.0
flask-cors>=3.0.10
tf-keras>=2.15.0
numpy<2.0
